+++ saleor/__init__.py
@@ -1,3 +1,5 @@
+import pillow_avif  # noqa: F401 # imported for side effects
+
 from .celeryconf import app as celery_app
 
 __all__ = ["celery_app"]